move CBImpl out of API
