+++ tt_metal/impl/program/dispatch.cpp
@@ -32,6 +32,7 @@
 #include <tt_stl/assert.hpp>
 #include "buffer.hpp"
 #include "circular_buffer.hpp"
+#include "impl/buffers/circular_buffer.hpp"
 #include "circular_buffer_constants.h"
 #include "core_coord.hpp"
 #include "device.hpp"
